Renaming file to match new name of project
