+++ spring-sync/src/main/java/com/springsync/springsync/MultiThreadedApisApplication.java
@@ -1,4 +1,4 @@
-package com.springsync.multithreadedapis;
+package com.springsync.springsync;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
