+++ spring-sync/src/main/java/com/springsync/springsync/controller/PlayerController.java
@@ -1,7 +1,7 @@
-package com.springsync.multithreadedapis.controller;
+package com.springsync.springsync.controller;
 import com.springsync.request.ThreadedPlayerRequest;
 import com.springsync.response.PlayerResponse;
-import com.springsync.multithreadedapis.services.PlayerService;
+import com.springsync.springsync.services.PlayerService;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
 import com.fasterxml.jackson.core.JsonProcessingException;
