+++ spring-sync/src/main/java/com/springsync/springsync/services/PlayerService.java
@@ -1,4 +1,4 @@
-package com.springsync.multithreadedapis.services;
+package com.springsync.springsync.services;
 import com.springsync.entity.Player;
 import com.springsync.entity.Sport;
 import com.springsync.entity.Team;
